requirements.txt
ccxt==2.9.96
pandas==2.0.3
numpy==1.24.3
python-telegram-bot==13.14
apscheduler==3.10.4